Use one file to Core classes
